extract `getConfig()` into config.ts
